back de creation sujet jusqu'a la liste questions

+++ src/main/java/repository/MatiereRepository.java
@@ -1,13 +1,15 @@
 package repository;
 
+import java.util.List;
+
 import org.springframework.data.repository.CrudRepository;
 
 import model.Matiere;
 import model.Theme;
 
 
-public interface MatiereRepository extends CrudRepository<Matiere, Long>{
+public interface MatiereRepository extends CrudRepository<Matiere, Integer>{
 	
 	
-
+	
 }
